requirements.txt
Django==1.8.1
Pillow==2.8.1
django-cors-headers==1.0.0
djangorestframework==3.1.2
wsgiref==0.1.2
